Move highlighted-annotation state from `selection` to `annotations`
Highlight state is relevant on an annotation-by-annotation basis but
does not have any impact on whether an annotation is present or
visible (at all) to the user, which is what `selection` state is for.

At this time, remove highlight-state logic from thread-building as this
is an extra complexity for something that isn't implemented in the UI
anyway.

In future, to re-implement highlight styling, query the `annotations`
store state (perhaps add a `annotationIsHighlighted` selector for
convenience).
