+++ src/sidebar/store/modules/selection.js
@@ -128,10 +128,6 @@ function init(settings) {
     // match the currently-applied filters
     forcedVisible: {},
 
-    // A map of annotations that should appear as "highlighted", e.g. the
-    // target of a single-annotation view
-    highlighted: {},
-
     filterQuery: settings.query || null,
     focusMode: setFocus(settings.focus || /** @type FocusConfig */ ({})),
 
@@ -209,10 +205,6 @@ const update = {
 
   SET_EXPANDED: function (state, action) {
     return { expanded: { ...state.expanded, [action.id]: action.expanded } };
-  },
-
-  HIGHLIGHT_ANNOTATIONS: function (state, action) {
-    return { highlighted: action.highlighted };
   },
 
   SELECT_TAB: function (state, action) {
@@ -342,23 +334,6 @@ function setExpanded(id, expanded) {
     type: actions.SET_EXPANDED,
     id,
     expanded,
-  };
-}
-
-/**
- * Highlight annotations with the given `ids`.
- *
- * This is used to indicate the specific annotation in a thread that was
- * linked to for example. Replaces the current map of highlighted annotations.
- * All provided annotations (`ids`) will be set to `true` in the `highlighted`
- * map.
- *
- * @param {string[]} ids - annotations to highlight
- */
-function highlightAnnotations(ids) {
-  return {
-    type: actions.HIGHLIGHT_ANNOTATIONS,
-    highlighted: toTrueMap(ids),
   };
 }
 
@@ -575,7 +550,6 @@ export default {
     clearSelectedAnnotations,
     clearSelection,
     focusAnnotations,
-    highlightAnnotations,
     selectAnnotations,
     selectTab,
     setExpanded,
