+++ src/sidebar/store/modules/test/selection-test.js
@@ -380,22 +380,6 @@ describe('sidebar/store/modules/selection', () => {
     });
   });
 
-  describe('highlightAnnotations()', function () {
-    it('sets the highlighted annotations', function () {
-      store.highlightAnnotations(['id1', 'id2']);
-      assert.deepEqual(getSelectionState().highlighted, {
-        id1: true,
-        id2: true,
-      });
-    });
-
-    it('replaces the current set of highlighted annotations', () => {
-      store.highlightAnnotations(['id1', 'id2']);
-      store.highlightAnnotations(['id3']);
-      assert.deepEqual(getSelectionState().highlighted, { id3: true });
-    });
-  });
-
   describe('isAnnotationFocused', () => {
     it('returns true if the provided annotation ID is in the set of focused annotations', () => {
       store.focusAnnotations([1, 2]);
