+++ src/sidebar/test/build-thread-test.js
@@ -489,27 +489,4 @@ describe('build-thread', function () {
       assert.deepEqual(thread[0].children[0].depth, 1);
     });
   });
-
-  describe('highlighting', function () {
-    it('does not set highlight state when none are highlighted', function () {
-      const thread = createThread(SIMPLE_FIXTURE, {}, ['dimmed']);
-      thread.forEach(function (child) {
-        assert.equal(child.highlightState, undefined);
-      });
-    });
-
-    it('highlights annotations', function () {
-      const thread = createThread(SIMPLE_FIXTURE, { highlighted: ['1'] }, [
-        'highlightState',
-      ]);
-      assert.equal(thread[0].highlightState, 'highlight');
-    });
-
-    it('dims annotations which are not highlighted', function () {
-      const thread = createThread(SIMPLE_FIXTURE, { highlighted: ['1'] }, [
-        'highlightState',
-      ]);
-      assert.equal(thread[1].highlightState, 'dim');
-    });
-  });
 });
